FEAT: complete most of the features and updating document

+++ src/components/comment/CommentDeleteBtn.tsx
@@ -5,24 +5,35 @@ import { useForm } from "react-hook-form";
 import { Button } from "../ui/button";
 import { deleteCommentAction } from "@/src/app/action/commentAction";
 import { useTranslations } from "next-intl";
+import { useSession } from "next-auth/react";
+import { useError } from "@/src/context/ErrorContext";
 const CommentDeleteBtn = ({
   comment_id,
   post_id,
   comments,
   onChangeEditState,
+  user_id,
 }: {
   comment_id: string;
   post_id: string;
   comments: number;
   onChangeEditState: (a: boolean) => void;
+  user_id: string;
 }) => {
   const {
     handleSubmit,
     formState: { isSubmitting },
   } = useForm();
+  const { setError } = useError();
+
   const onSubmit = async () => {
+    if (session!.user.id !== user_id) {
+      setError(new Error("Error test"));
+      return;
+    }
     await deleteCommentAction(comment_id, post_id, comments);
   };
+  const { data: session } = useSession();
   const t = useTranslations("Comment");
   return (
     <div className="flex gap-2">
